+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version 1
 ADDED:
 
 - new CHANGELOG.md file created
+- the icons from the RemixIcon project
 
 Notable links:
 [keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into changelogs